pyproject.toml
[project]
name = "hw-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "gradio>=5.47.2",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.6",
    "numpy<2.0",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "tensorflow==2.16.2",
    "tensorflow-metal>=1.2.0",
]
